modified classes and updated test case to follow code conventions
