+++ core-java-modules/core-java-lang-5/src/main/java/com/baeldung/shallowdeepcopy/AddressDeepCopy.java
@@ -1,0 +1,13 @@
+package com.baeldung.shallowdeepcopy;
+
+class AddressDeepCopy {
+    String city;
+
+    AddressDeepCopy(String city) {
+        this.city = city;
+    }
+
+    protected AddressDeepCopy clone() {
+        return new AddressDeepCopy(this.city);
+    }
+}
