+++ core-java-modules/core-java-lang-5/src/main/java/com/baeldung/shallowdeepcopy/AddressShallowCopy.java
@@ -1,0 +1,9 @@
+package com.baeldung.shallowdeepcopy;
+
+class AddressShallowCopy {
+    String city;
+
+    AddressShallowCopy(String city) {
+        this.city = city;
+    }
+}
