+++ core-java-modules/core-java-lang-5/src/main/java/com/baeldung/shallowdeepcopy/PersonDeepCopy.java
@@ -1,18 +1,5 @@
 package com.baeldung.shallowdeepcopy;
 
-class AddressDeepCopy {
-    String city;
-
-    AddressDeepCopy(String city) {
-        this.city = city;
-    }
-
-    protected AddressDeepCopy clone() {
-        // Strings are immutable in Java, so copying them (deep or shallow) results in new objects
-        return new AddressDeepCopy(this.city);
-    }
-}
-
 class PersonDeepCopy implements Cloneable {
     String name;
     AddressDeepCopy address;
@@ -22,7 +9,6 @@ class PersonDeepCopy implements Cloneable {
         this.address = address;
     }
 
-    // Creates a deep copy of the Person object
     @Override
     protected PersonDeepCopy clone() {
         return new PersonDeepCopy(this.name, this.address.clone());
